move isFrontCameraAvailable back to CaptureSelfieActivity classes https://github.com/opendatakit/collect/pull/1985#issuecomment-371984438
